+++ lib/install.ps1
@@ -1,4 +1,4 @@
-'Helpers', 'autoupdate', 'buckets', 'decompress', 'manifest' | ForEach-Object {
+'Helpers', 'autoupdate', 'buckets', 'decompress', 'manifest', 'ManifestHelpers' | ForEach-Object {
     . (Join-Path $PSScriptRoot "$_.ps1")
 }